+++ docs/releases.md
@@ -3,6 +3,11 @@
 This page will give you the change that are occuring when a new version has been published on pypi.
 The changes have been tracked starting version 0.1.0
 
+## version 0.3.0
+
+* Fix an issue where segments were not considered in findComponentUsage when they were used in columns
+* adding `RequestCreator` class
+
 ## version 0.2.9
 
 * Adding the `getScheduledJobs` endpoint